chore: Component
Finished the SVG creation.

Component : Started creating '#component' element and all elements inside. Created logo and the search bar with two buttons (one for searching a task and the other to create a task), and the filters bar with 3 buttons inside (one filters to see all tasks, the second only tasks in progress, and the last to see only finished tasks).