chore: JS, HTML and Style
Component : Finished to place filter buttons, added an event listener (src/assets/scripts/events.js) to them (will add a class to the clicked one, which is changing its color and will filter tasks if they are not finished, finished or all). Stylized the buttons and the input.

Added the 'Montserrat' font.