chore: Events
dragndrop.js : when dragndropping a task, it changes the order in the localstorage item to load them in this order when loading the page.

events.js : added a click event on the modal form to make it disapear when clicking outside itself, added a keydown event on the search input to search a task, added a click event on finished buttons to change style of a task and make it clear that it's finished, starting the filter option with a click event to filter task to display finished, not finished tasks or display both.

elements.js : change the classlist property in the button element to make it manage arrays (in order to add multiple class to a same element)

index.html : added template element to replace the js createelement template.