chore: Script (events, task appening) and Style (dark/light theme, scrolling task box)
<JavaScript>

add.js, function.js : Browser's Local Storage is now used to store stringified JSON of the inner HTML of the Task container in order to make it easier for the dev to store all tasks in one time.

events.js : Finished the filter option, so now we can filter our tasks by clicking the buttons above them (All Tasks, Not Finished and Finished). Added a button to switch dark/light theme.

var.js : Changed the class and put all the properties in a method in order to refresh them when we need that.

<Style>

media-color.scss : Styled the prefered color scheme dark and light with the 'theme-b' HTML class to switch between both.

modal-add.scss : Styled invalid input and added a label in order to show the error to users.

tasks-section.scss : Styled the Task container, we can now scroll into it if we have too many tasks to display.