chore: Style and Media Queries
<JavaScript>

add.js, event.js : Changed the Options button click events to apply them also on the mobile one (appears in the footer).
footer.js : Created the footer (will appear only on mobile devices, 425px wide max) and the buttons inside.

<Style>

Styled the footer, the scrollbar in the taskbox, the bins in the tasks, and modified media queries 768px and 425px wide max.