chore: Script, Style
<JavaScript>
Deleted var.js, which contained all the querySelectors I needed in multiple files, I simply replaced them by querySelector but inside the script.
Some semicolons added at end of some lines.
add.js : Correction of a bug with the local storage data gathering, if there wasn't any 'task' key in the local storage, the script ended there. Now, in the local storage I set a string but no more object to avoid this. And because an object was useless in my case.

<Style>
medias-color.scss : Background are changed when the media is maximum 768px wide.
medias-size.scss : The taskbox is a bit taller when the medias is max 768px wide.