chore: Sass function, Checked event
<JavaScript>
event.js : Changed the click event on the checkbox to put it on the task. It will change the task class when the event target is the checkbox or the text.

<Style>
variables.scss : Added a function that decrease the alpha channel of a color. Used it on the tasks text.