chore: Fixed Drag and Drop, Edit button
<JavaScript>
dragndrop.js : Fixed the dragndrop drop event, it fired twice, stopped it with the 'Event.stopImmediatePropagation()' method.
events.js : The event for checking tasks is now on the task itself, for a better experience on tactile devices. Added an edit icon : can modify the task name by clicking it.
functions.js : 'refreshEvents()' function created in ordrer to refresh those events in less code lines for a better reading.