chore: Local storage
<JavaScript>
add.js, function.js : Changed the storeData() function, now it stores a JSON object with the tasks text ('name' property) and boolean ('checked' property) that determine if the tasks was marked as completed or not.
event.js : Some events was fired twice (or more), added the Event.stopImmediatePropagation() method to avoid this.

<Style>
medias-size.scss : Changed the modal form elements' width for the 425px media query, in order to make it fit the screen width.