fixes: General fixes
<JavaScript>
Commented the most part of the code for a better reading.
add.js : Added a condition in the task creation event in order to make the new task disapear when the user has the 'Finished' filter active. Moved the theme storing from index.js to add.js.
event.js : Changed the localStorage.clear() (line 22) to localStorage.removeItem('taskList') to make it not delete the stored color theme.
function.js : Deleted the timeout method, can't clear it when the modal closes. The hint message will stay visible util the modal window is open.

<Style>
Commented and rearranged the most part of the code for a better reading.